Use literal newlines in status
